+++ src/r3f/components/TilesRenderer.jsx
@@ -1,9 +1,11 @@
-import { createContext, useContext, useEffect, useRef, forwardRef, useMemo, useCallback } from 'react';
+import { createContext, useContext, useEffect, useRef, forwardRef, useMemo, useCallback, useState } from 'react';
 import { useThree, useFrame } from '@react-three/fiber';
+import { Object3D } from 'three';
 import { TilesRenderer as TilesRendererImpl } from '../../three/TilesRenderer.js';
 import { useDeepOptions, useShallowOptions } from '../utilities/useOptions.js';
 import { useObjectDep } from '../utilities/useObjectDep.js';
 import { useApplyRefs } from '../utilities/useApplyRefs.js';
+import { WGS84_ELLIPSOID } from '../../three/math/GeoConstants.js';
 
 // context for accessing the tile set
 export const TilesRendererContext = createContext( null );
@@ -36,31 +38,78 @@ export function EastNorthUpFrame( props ) {
 		az = 0,
 		el = 0,
 		roll = 0,
+		ellipsoid = WGS84_ELLIPSOID.clone(),
 		children,
 	} = props;
-	const ref = useRef();
 	const tiles = useContext( TilesRendererContext );
 	const invalidate = useThree( state => state.invalidate );
+	const [ group, setGroup ] = useState( null );
 	const updateCallback = useCallback( () => {
 
-		// hide the group if the tiles aren't loaded yet
-		const ellipsoid = tiles && tiles.ellipsoid || null;
-		const group = ref.current;
-		group.matrix.identity();
-		group.visible = Boolean( tiles && tiles.root );
-
-		if ( ellipsoid === null ) {
+		if ( group === null ) {
 
 			return;
 
 		}
 
-		ellipsoid.getFrame( lat, lon, az, el, roll, height, group.matrix );
+		// hide the group if the tiles aren't loaded yet
+		const localEllipsoid = tiles && tiles.ellipsoid || ellipsoid || null;
+		group.matrix.identity();
+		group.visible = Boolean( tiles && tiles.root || ellipsoid );
+
+		if ( localEllipsoid === null ) {
+
+			return;
+
+		}
+
+		localEllipsoid.getFrame( lat, lon, az, el, roll, height, group.matrix );
 		group.matrix.decompose( group.position, group.quaternion, group.scale );
 		group.updateMatrixWorld();
 		invalidate();
 
-	}, [ invalidate, tiles, lat, lon, height, az, el, roll ] );
+	}, [ invalidate, tiles, lat, lon, height, az, el, roll, ellipsoid, group, useObjectDep( ellipsoid.radius ) ] ); // eslint-disable-line react-hooks/exhaustive-deps
+
+	// adjust the matrix world update logic if a tile set is present so that we can position the frame
+	// correctly regardless of the parent.
+	useEffect( () => {
+
+		if ( tiles !== null && group !== null ) {
+
+			group.updateMatrixWorld = function ( force ) {
+
+				if ( this.matrixAutoUpdate ) {
+
+					this.updateMatrix();
+
+				}
+
+				if ( this.matrixWorldNeedsUpdate || force ) {
+
+					this.matrixWorld.multiplyMatrices( tiles.group.matrixWorld, this.matrix );
+					force = true;
+
+				}
+
+				const children = this.children;
+				for ( let i = 0, l = children.length; i < l; i ++ ) {
+
+					const child = children[ i ];
+					child.updateMatrixWorld( force );
+
+				}
+
+			};
+
+			return () => {
+
+				group.updateMatrixWorld = Object3D.prototype.updateMatrixWorld;
+
+			};
+
+		}
+
+	}, [ tiles, group ] );
 
 	useEffect( () => {
 
@@ -68,7 +117,14 @@ export function EastNorthUpFrame( props ) {
 
 	}, [ updateCallback ] );
 
+	// update the position when a tile set is loaded since it may modify the ellipsoid
 	useEffect( () => {
+
+		if ( tiles === null ) {
+
+			return;
+
+		}
 
 		tiles.addEventListener( 'load-tile-set', updateCallback );
 		return () => {
@@ -79,7 +135,7 @@ export function EastNorthUpFrame( props ) {
 
 	}, [ tiles, updateCallback ] );
 
-	return <group ref={ ref }>{ children }</group>;
+	return <group ref={ setGroup }>{ children }</group>;
 
 }