+++ CHANGELOG.md
@@ -5,9 +5,33 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- TilesRenderer: No longer logs a warning if no cameras are present and a custom plugin supports tile error.
+- Type definition errors.
+- R3F EastNorthUpFrame: frame will be positioned correctly regardless of parent.
+- TilesRenderer: Fixed screen space error being calculated incorrectly with multiple cameras.
+
+### Added
+- R3F CompassGizmo: Support for X & Y margin for CompassGizmo.
+- CameraTransitionManager: Expose "alpha" via the change event and class member.
+
+## [0.4.7] - 2025.03.03
 ### Added
 - GlobeControls: If no raycast intersection is fund then fallback to intersecting the tile set globe.
 - EnvironmentControls, GlobeControls: Fix inertia calculations for orthographic cameras.
+- R3F EastNorthUpFrame: Add support for passing the ellipsoid in directly.
+- TilesRenderer.group: Added "matrixWorldInverse" field.
+- Add "LoadRegionPlugin".
+
+### Fixed
+- Fix case where the environment and globe controls can have residual inertia after stopping the mouse.
+- Image Format Plugins: Fix case where tiles may not cause update on first load resulting in no tiles rendered.
+
+### Changed
+- Deprecated "TilesRenderer.setLatLonToYUp" function.
+- Deprecated "TilesRenderer.errorThreshold" option.
+- Deprecated "TilesRenderer.optimizeRaycast" option.
+- Use "TilesRenderer.group.matrixWorldInverse" across src files to reduce matrix invert operations.
 
 ## [0.4.6] - 2025.02.21
 ### Added